+++ EhPanda/Database/MangaStateMO+CoreDataProperties.swift
@@ -15,11 +15,11 @@ extension MangaStateMO: GalleryIdentifiable {
     @NSManaged public var aspectBox: Data?
     @NSManaged public var comments: Data?
     @NSManaged public var contents: Data?
-    @NSManaged public var currentPageNum: Int16
+    @NSManaged public var currentPageNum: Int64
     @NSManaged public var gid: String
-    @NSManaged public var pageNumMaximum: Int16
+    @NSManaged public var pageNumMaximum: Int64
     @NSManaged public var previews: Data?
-    @NSManaged public var readingProgress: Int16
+    @NSManaged public var readingProgress: Int64
     @NSManaged public var tags: Data?
     @NSManaged public var userRating: Float
 }
